JohnCMS 3.0.0
[new]   Серьезное обновление ядра системы, по-новому организованы настройки
[new]   Новая, быстродействующая система подсчета юзеров/гостей онлайн
[new]   Новый Форум! Быстродействующее ядро, оптимизированные SQL запросы
        - голосования в темах
        - фильтрация постов по авторам
        - автоподстановка названий тем форума, в ссылки
        - прикрепление первого поста в теме (отключается из личных настроек)
        - полнотекстовый поиск с сортировкой по релевантности запросов
        - подсветка PHP кода в текстах (BBcode)
        - настройка "обрезки" длинных постов
        - возможность для юзера редактировать/удалять свой последний пост
        - много других улучшений
[new]   Новый модуль восстановления забытого пароля
[new]   Установка прав доступа к модулям (все, авторизованные, только админы)
[new]   Файл-менеджер (обменник) для прикрепленных файлов форума
        - Сортировка по типу файлов (архивы, картинки, тексты и т.д.)
        - Фильтрация по разделу, подразделу, топику
        - Список "новых" файлов форума
[new]   Новая админка
        - Удаление юзеров с выборочной чисткой активности
        - Расширенный поиск по Нику, или адресу IP (маска, диапазон)
        - Удобное управление Форумом
        - Много других новых функций
[new]   Новый загруз-центр, быстрый алгоритм, оптимизированные SQL запросы
[new]   Продвинутый поиск пользователей
[new]   Новая функция парсинга BBcode для ссылок
[new]   Новая (быстрая, с Кэшированием) функция парсинга Cмайлов и их каталог
[new]   Дайджест показывает события для Администратора сайта (события в Админке)
[new]   Блокировка одинаковых сообщений
[new]   "Монитор активности", переход к последнему посту из списка последних 10
[new]   Топ активности посетителей Сайта
[new]   Новый "личный кабинет"
[new]   Новая анкета. По-новому организовано ее редактирование
[new]   Модуль управления рекламными ссылками
[new]   Выгрузка фото в личную Анкету
[new]   Система Аватаров
[new]   Тема оформления от `oz`
[mod]   Оптимизированы функции статистики, значительно увеличена скорость работы
[mod]   Оптимизирован код Гостевой
[mod]   Оптимизирован код списка "Онлайн", снижена нагрузка на сервер
[mod]   Возможность редактировать статью и ее параметры в Библиотеке
[mod]   Администратор может писать даже в закрытые темы форума
[mod]   Библиотека в заголовке страницы показывает название раздела / статьи
[mod]   Применена новая функция pagenav() при чтении статьи в Библиотеке
[fix]   Удаленные юзеры теперь помечаются специальными значками
[fix]   При редактировании новости, могла возникать ошибка XHTML
[fix]   Устранены ошибки в проверке CAPTCHA
[fix]   Исключены ошибки XHTML, которые возникали на заголовках Библиотеки
[fix]   Устранена ошибка в модуле HTTP antiFlood

+++ str/news.php
@@ -21,113 +21,113 @@ $headmod = "news";
 require_once ("../incfiles/core.php");
 require_once ("../incfiles/head.php");
 
-$do = isset($_GET['do']) ? $_GET['do'] : '';
-switch ($do)
-{
-    case 'add':
-        ////////////////////////////////////////////////////////////
-        // Добавление новости                                     //
-        ////////////////////////////////////////////////////////////
-        if ($dostsmod == 1)
-        {
-            echo '<div class="phdr">Добавить новость</div>';
-            $old = 20;
-            if ($lastpost > ($realtime - $old))
-            {
-                echo '<p><b>Антифлуд!</b><br />Вы не можете так часто писать<br/>Порог ' . $old . ' секунд<br/><br/><a href="news.php">Назад</a></p>';
-                require_once ("../incfiles/end.php");
-                exit;
-            }
-            if (isset($_POST['submit']))
-            {
-
-                if (empty($_POST['name']))
-                {
-                    echo "Вы не ввели заголовок<br/><a href='news.php?act=new'>Повторить</a><br/>";
+$do
+    = isset ($_GET['do']) ? $_GET['do'] : '';
+switch ($do
+        ) {
+        case 'add' :
+            ////////////////////////////////////////////////////////////
+            // Добавление новости                                     //
+            ////////////////////////////////////////////////////////////
+            if ($rights >= 6) {
+                echo '<div class="phdr">Добавить новость</div>';
+                $old = 20;
+                if ($datauser['lastpost'] > ($realtime - $old)) {
+                    echo '<p><b>Антифлуд!</b><br />Вы не можете так часто писать<br/>Порог ' . $old . ' секунд<br/><br/><a href="news.php">Назад</a></p>';
                     require_once ("../incfiles/end.php");
                     exit;
                 }
-                if (empty($_POST['text']))
-                {
-                    echo "Вы не ввели текст<br/><a href='news.php?act=new'>Повторить</a><br/>";
-                    require_once ("../incfiles/end.php");
-                    exit;
-                }
-                $name = check($_POST['name']);
-                $text = mysql_real_escape_string(trim($_POST['text']));
-                if (!empty($_POST['pf']) && ($_POST['pf'] != '0'))
-                {
-                    $pf = intval($_POST['pf']);
-                    $rz = $_POST['rz'];
-                    $pr = mysql_query("select * from `forum` where type='r' and refid= '" . $pf . "';");
-                    while ($pr1 = mysql_fetch_array($pr))
-                    {
-                        $arr[] = $pr1['id'];
+                if (isset ($_POST['submit'])) {
+                    if (empty ($_POST['name'])) {
+                        echo "Вы не ввели заголовок<br/><a href='news.php?act=new'>Повторить</a><br/>";
+                        require_once ("../incfiles/end.php");
+                        exit;
                     }
-                    foreach ($rz as $v)
-                    {
-                        if (in_array($v, $arr))
-                        {
-                            mysql_query("insert into `forum` values(0,'" . $v . "','t','" . $realtime . "','" . $login . "','','','','','" . $name . "','','','1','','','','','');");
-                            $tem = mysql_query("select * from `forum` where type='t' and time='" . $realtime . "' and refid= '" . $v . "';");
-                            $tem1 = mysql_fetch_array($tem);
-                            $agn = strtok($agn, ' ');
-                            mysql_query("insert into `forum` values(0,'" . $tem1['id'] . "','m','" . $realtime . "','" . $login . "','','','" . $ipp . "','" . $agn . "','" . $text . "','','','','','','','','');");
+                    if (empty ($_POST['text'])) {
+                        echo "Вы не ввели текст<br/><a href='news.php?act=new'>Повторить</a><br/>";
+                        require_once ("../incfiles/end.php");
+                        exit;
+                    }
+                    $name = check($_POST['name']);
+                    $text = trim($_POST['text']);
+                    if (!empty ($_POST['pf']) && ($_POST['pf'] != '0')) {
+                        $pf = intval($_POST['pf']);
+                        $rz = $_POST['rz'];
+                        $pr = mysql_query("SELECT * FROM `forum` WHERE `refid` = '$pf' AND `type` = 'r'");
+                        while ($pr1 = mysql_fetch_array($pr)) {
+                            $arr[] = $pr1['id'];
+                        }
+                        foreach ($rz as $v) {
+                            if (in_array($v, $arr)) {
+                                mysql_query(
+                                "INSERT INTO `forum` SET
+                            `refid` = '$v',
+                            `type` = 't',
+                            `time` = '$realtime',
+                            `user_id` = '$user_id',
+                            `from` = '$login',
+                            `text` = '$name'"
+                                );
+                                $rid = mysql_insert_id();
+                                mysql_query(
+                                "INSERT INTO `forum` SET
+                            `refid` = '$rid',
+                            `type` = 'm',
+                            `time` = '$realtime',
+                            `user_id` = '$user_id',
+                            `from` = '$login',
+                            `ip` = '$ipp',
+                            `soft` = '"
+                                . mysql_real_escape_string($agn) . "',
+                            `text` = '" . mysql_real_escape_string($text) . "'");
+                            }
                         }
                     }
+                    mysql_query("insert into `news` values(0,'$realtime','$login','$name','" . mysql_real_escape_string($text) . "','$rid')");
+                    mysql_query("UPDATE `users` SET `lastpost` = '$realtime' WHERE `id` = '$user_id'");
+                    echo "Новость добавлена.<p><a href='news.php'>Продолжить</a></p>";
                 }
-                mysql_query("insert into `news` values(0,'" . $realtime . "','" . $login . "','" . $name . "','" . $text . "','" . $tem1[id] . "');");
-                mysql_query("UPDATE `users` SET `lastpost` = '" . $realtime . "' WHERE `id` = '" . $user_id . "'");
-                echo "Новость добавлена.<p><a href='news.php'>Продолжить</a></p>";
-            } else
-            {
-                echo '<form action="news.php?do=add" method="post">';
-                echo '<div class="menu"><u>Заголовок</u><br/><input type="text" name="name"/></div>';
-                echo '<div class="menu"><u>Текст</u><br/><textarea rows="4" name="text"></textarea></div>';
-                echo '<div class="menu"><u>Раздел форума для обсуждения новости</u><br/>';
-                $fr = mysql_query("SELECT * FROM `forum` WHERE `type` = 'f'");
-                echo '<input type="radio" name="pf" value="0" checked="checked" />Не обсуждать<br />';
-                while ($fr1 = mysql_fetch_array($fr))
-                {
-                    echo "<input type='radio' name='pf' value='" . $fr1['id'] . "'/>$fr1[text]<select name='rz[]'>";
-                    $pr = mysql_query("select * from `forum` where type='r' and refid= '" . $fr1['id'] . "';");
-                    while ($pr1 = mysql_fetch_array($pr))
-                    {
-                        echo '<option value="' . $pr1['id'] . '">' . $pr1['text'] . '</option>';
+                else {
+                    echo '<form action="news.php?do=add" method="post">';
+                    echo '<div class="menu"><u>Заголовок</u><br/><input type="text" name="name"/></div>';
+                    echo '<div class="menu"><u>Текст</u><br/><textarea rows="4" name="text"></textarea></div>';
+                    echo '<div class="menu"><u>Раздел форума для обсуждения новости</u><br/>';
+                    $fr = mysql_query("SELECT * FROM `forum` WHERE `type` = 'f'");
+                    echo '<input type="radio" name="pf" value="0" checked="checked" />Не обсуждать<br />';
+                    while ($fr1 = mysql_fetch_array($fr)) {
+                        echo "<input type='radio' name='pf' value='" . $fr1['id'] . "'/>$fr1[text]<select name='rz[]'>";
+                        $pr = mysql_query("select * from `forum` where type='r' and refid= '" . $fr1['id'] . "'");
+                        while ($pr1 = mysql_fetch_array($pr)) {
+                            echo '<option value="' . $pr1['id'] . '">' . $pr1['text'] . '</option>';
+                        }
+                        echo '</select><br/>';
                     }
-                    echo '</select><br/>';
+                    echo '</div><div class="bmenu"><input type="submit" name="submit" value="Ok!"/></div></form><p><a href="news.php">К новостям</a></p>';
                 }
-                echo '</div><div class="bmenu"><input type="submit" name="submit" value="Ok!"/></div></form><p><a href="news.php">К новостям</a></p>';
             }
-        } else
-        {
-            header("location: news.php");
-        }
-        break;
+            else {
+                header("location: news.php");
+            }
+            break;
 
-    case 'edit':
+        case 'edit' :
         ////////////////////////////////////////////////////////////
         // Редактирование новости                                 //
         ////////////////////////////////////////////////////////////
-        if ($dostsmod == 1)
-        {
+        if ($rights >= 6) {
             echo '<div class="phdr">Редактирование новости</div>';
-            if (empty($_GET['id']))
-            {
+            if (empty ($_GET['id'])) {
                 echo "Ошибка!<br/><a href='news.php'>К новостям</a><br>";
                 require_once ("../incfiles/end.php");
                 exit;
             }
-            if (isset($_POST['submit']))
-            {
-                if (empty($_POST['name']))
-                {
+            if (isset ($_POST['submit'])) {
+                if (empty ($_POST['name'])) {
                     echo "Вы не ввели заголовок<br/><a href='news.php?act=edit&amp;id=" . $id . "'>Повторить</a><br/>";
                     require_once ("../incfiles/end.php");
                     exit;
                 }
-                if (empty($_POST['text']))
-                {
+                if (empty ($_POST['text'])) {
                     echo "Вы не ввели текст<br/><a href='news.php?act=edit&amp;id=" . $id . "'>Повторить</a><br/>";
                     require_once ("../incfiles/end.php");
                     exit;
@@ -139,52 +139,49 @@ switch ($do)
 			`text` = '" . $text . "'
 			WHERE `id` = '" . $id . "';");
                 echo '<p>Новость изменена.<br /><a href="news.php">Продолжить</a></p>';
-            } else
-            {
-                $req = mysql_query("select * from `news` where id='" . $id . "';");
-                $res = mysql_fetch_array($req);
-                echo '<form action="news.php?do=edit&amp;id=' . $id . '" method="post">Заголовок:<br/><input type="text" name="name" value="' . $res['name'] . '"/><br/>Текст:<br/><textarea cols="30" rows="5" name="text">' . $res['text'] .
-                    '</textarea><br/><input type="submit" name="submit" value="Ok!"/></form><p><a href="news.php">К новостям</a></p>';
             }
-        } else
-        {
+            else {
+                $req = mysql_query("SELECT * FROM `news` WHERE `id` = '" . $id . "'");
+                $res = mysql_fetch_array($req);
+                echo '<form action="news.php?do=edit&amp;id=' . $id . '" method="post">Заголовок:<br/><input type="text" name="name" value="' . $res['name'] . '"/><br/>Текст:<br/><textarea cols="30" rows="5" name="text">' .
+                htmlentities($res['text'], ENT_QUOTES, 'UTF-8') . '</textarea><br/><input type="submit" name="submit" value="Ok!"/></form><p><a href="news.php">К новостям</a></p>';
+            }
+        }
+        else {
             header("location: news.php");
         }
         break;
 
-    case 'clean':
+    case 'clean' :
         ////////////////////////////////////////////////////////////
         // Чистка новостей                                        //
         ////////////////////////////////////////////////////////////
-        if ($dostadm == 1)
-        {
+        if ($rights >= 7) {
             echo '<div class="phdr">Чистка новостей</div>';
-            if (isset($_POST['submit']))
-            {
-                $cl = isset($_POST['cl']) ? intval($_POST['cl']) : '';
-                switch ($cl)
-                {
-                    case '1':
+            if (isset ($_POST['submit'])) {
+                $cl = isset ($_POST['cl']) ? intval($_POST['cl']) : '';
+                switch ($cl) {
+                    case '1' :
                         // Чистим новости, старше 1 недели
                         mysql_query("DELETE FROM `news` WHERE `time`<='" . ($realtime - 604800) . "'");
                         mysql_query("OPTIMIZE TABLE `news`;");
                         echo '<p>Удалены все новости, старше 1 дня.</p><p><a href="news.php">К новостям</a></p>';
                         break;
 
-                    case '2':
+                    case '2' :
                         // Проводим полную очистку
                         mysql_query("TRUNCATE TABLE `news`");
                         echo '<p>Удалены все новости.</p><p><a href="news.php">К новостям</a></p>';
                         break;
 
-                    default:
+                    default :
                         // Чистим сообщения, старше 1 месяца
                         mysql_query("DELETE FROM `news` WHERE `time`<='" . ($realtime - 2592000) . "'");
                         mysql_query("OPTIMIZE TABLE `news`;");
                         echo '<p>Удалены все новости, старше 1 недели.</p><p><a href="news.php">К новостям</a></p>';
                 }
-            } else
-            {
+            }
+            else {
                 echo '<p><u>Что чистим?</u>';
                 echo '<form id="clean" method="post" action="news.php?do=clean">';
                 echo '<input type="radio" name="cl" value="0" checked="checked" />Старше 1 месяца<br />';
@@ -194,34 +191,32 @@ switch ($do)
                 echo '</form></p>';
                 echo '<p><a href="news.php">Отмена</a></p>';
             }
-        } else
-        {
+        }
+        else {
             header("location: news.php");
         }
         break;
 
-    case 'del':
+    case 'del' :
         ////////////////////////////////////////////////////////////
         // Удаление новости                                       //
         ////////////////////////////////////////////////////////////
-        if ($dostsmod == 1)
-        {
+        if ($rights >= 6) {
             echo '<div class="phdr">Удалить новость</div>';
-            if (isset($_GET['yes']))
-            {
+            if (isset ($_GET['yes'])) {
                 mysql_query("DELETE FROM `news` WHERE `id` = '" . $id . "' LIMIT 1");
                 echo '<p>Новость удалена!<br/><a href="news.php">К новостям</a></p>';
-            } else
-            {
+            }
+            else {
                 echo '<p>Вы уверены,что хотите удалить новость?<br/><a href="news.php?do=del&amp;id=' . $id . '&amp;yes">Да</a> | <a href="news.php">Нет</a></p>';
             }
-        } else
-        {
+        }
+        else {
             header("location: news.php");
         }
         break;
 
-    default:
+    default :
         ////////////////////////////////////////////////////////////
         // Вывод списка новостей                                  //
         ////////////////////////////////////////////////////////////
@@ -229,31 +224,24 @@ switch ($do)
         $req = mysql_query("SELECT COUNT(*) FROM `news`");
         $total = mysql_result($req, 0);
         $req = mysql_query("SELECT * FROM `news` ORDER BY `time` DESC LIMIT " . $start . "," . $kmess . ";");
-        while ($nw1 = mysql_fetch_array($req))
-        {
-            echo ceil(ceil($i / 2) - ($i / 2)) == 0 ? '<div class="list1">' : '<div class="list2">';
+        while ($nw1 = mysql_fetch_array($req)) {
+            echo is_integer($i / 2) ? '<div class="list1">' : '<div class="list2">';
             $text = $nw1['text'];
             $text = htmlentities($text, ENT_QUOTES, 'UTF-8');
-			$text = str_replace("\r\n", "<br/>", $text);
+            $text = str_replace("\r\n", "<br/>", $text);
             $text = tags($text);
-            if ($offsm != 1 && $offgr != 1)
-            {
-                $text = smiles($text);
-                $text = smilescat($text);
-                $text = smilesadm($text);
-            }
-            $vr = $nw1['time'] + $sdvig * 3600;
+            if ($set_user['smileys'])
+                $text = smileys($text, 1);
+            $vr = $nw1['time'] + $set_user['sdvig'] * 3600;
             $vr1 = date("d.m.y / H:i", $vr);
             echo '<b>' . $nw1['name'] . '</b><br/>' . $text . '<div class="func"><font color="#999999">Добавил: ' . $nw1['avt'] . ' (' . $vr1 . ')</font><br/>';
-            if ($nw1['kom'] != 0 && $nw1['kom'] != "")
-            {
+            if ($nw1['kom'] != 0 && $nw1['kom'] != "") {
                 $mes = mysql_query("SELECT COUNT(*) FROM `forum` WHERE `type` = 'm' AND `refid` = '" . $nw1['kom'] . "'");
                 $komm = mysql_result($mes, 0) - 1;
                 if ($komm >= 0)
                     echo '<a href="../forum/?id=' . $nw1['kom'] . '">Обсудить на форуме (' . $komm . ')</a><br/>';
             }
-            if ($dostsmod == 1)
-            {
+            if ($rights >= 6) {
                 echo '<a href="news.php?do=edit&amp;id=' . $nw1['id'] . '">Изменить</a> | <a href="news.php?do=del&amp;id=' . $nw1['id'] . '">Удалить</a>';
             }
             echo '</div></div>';
@@ -261,16 +249,14 @@ switch ($do)
         }
         echo '<div class="phdr">Всего:&nbsp;' . $total . '</div>';
         echo '<p>';
-        if ($total > $kmess)
-        {
+        if ($total > $kmess) {
             echo '<p>' . pagenav('news.php?', $start, $total, $kmess) . '</p>';
             echo '<p><form action="index.php" method="get"><input type="hidden" name="act" value="new"/><input type="text" name="page" size="2"/><input type="submit" value="К странице &gt;&gt;"/></form></p>';
         }
         echo '</p>';
-        if ($dostsmod == 1)
-        {
+        if ($rights >= 6) {
             echo '<p><a href="news.php?do=add">Добавить новость</a>';
-            if ($dostadm == 1)
+            if ($rights >= 7)
                 echo '<br /><a href="news.php?do=clean">Чистка новостей</a>';
             echo '</p>';
         }
